chore(): improve comments from PR, add util file in bal-input
